Make polyfit2d() compatible with numpy.polynomial
Previously, the 2-dim polynomial fitting routine in this package
returned polynomial coefficients as an 1-dim array.  However, since its
version 1.7.0, NumPy added functions that evaluate 2-dim polynomials
(`numpy.polynomial.polyval2d()`, etc.).  Those functions expect that
polynomial coefficients are given as a 2-dim array, satisfying::

    p(x, y) = sum[i=0:n](sum[j=0:n] c_ij * x^i * y^j)

Thus, the polynomial coefficients returned by the fitting routine have
been incompatible with the relevant NumPy functions.

To fix the incompatibility, I implemented a 2-dim polynomial fitting
routine `polyfit2d()`.  The function assumes the 2-dim polynomial in a
form expressed as::

    p(x, y) = sum[i=0:n](sum[j=0:n] c_ij * x^i * y^j)

with a constraint of ``i + j <= n`` and returns the polynomial
coefficients `c` (`c_ij`) as a 2-dim array.  Note that, due to the
constraint, the elements in the lower right triangle of `c` are zeros.