Drop an extension module _poly2d
The extension module `_poly2d` was originally introduced to accelerate
computation of 2-dim polynomials. The functions in the module expect the
coefficients of a polynomial in an 1-dim array.  However, that form is
incompatible with newly introduced `polyval2d()`, which expects the
coefficients in a 2-dim array.  In addition, I suspected that the
benefit of the acceleration is marginal.  Thus, I have decided to
entirely drop `_poly2d`.